--- showhive-boot/admin/src/main/java/com/showhive/admin/application/command/dto/ListSeatGradeDto.java
@@ -1,16 +1,0 @@
-package com.showhive.admin.application.command.dto;
-
-import com.showhive.admin.interfaces.performance.dto.SeatGradeRequest;
-import lombok.Builder;
-
-@Builder
-public record ListSeatGradeDto(
-        Long seatGradeId,
-        String grade
-) {
-    public static ListSeatGradeDto of(SeatGradeRequest request) {
-        return ListSeatGradeDto.builder()
-                .grade(request.name())
-                .build();
-    }
-}
