--- showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/seatgrade/CreateSeatGradeUseCase.java
@@ -1,7 +1,0 @@
-package com.showhive.admin.application.command.usecase.seatgrade;
-
-import com.showhive.admin.application.command.dto.SeatGradeDto;
-
-public interface CreateSeatGradeUseCase {
-    void handle(SeatGradeDto commandDto);
-}
