--- showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/seatgrade/ReadSeatGradeUseCase.java
@@ -1,11 +1,0 @@
-package com.showhive.admin.application.command.usecase.seatgrade;
-
-import com.showhive.admin.interfaces.performance.dto.SeatGradeListResponse;
-import com.showhive.admin.interfaces.performance.dto.SeatGradeResponse;
-
-public interface ReadSeatGradeUseCase {
-
-    SeatGradeListResponse readAllSeatGrades(int pageSize, long lastGradeId);
-
-    SeatGradeResponse readSeatGrade(long seatGradeId);
-}
