--- showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/seatgrade/UpdateSeatGradeUseCase.java
@@ -1,7 +1,0 @@
-package com.showhive.admin.application.command.usecase.seatgrade;
-
-import com.showhive.admin.application.command.dto.SeatGradeDto;
-
-public interface UpdateSeatGradeUseCase {
-    void update(long seatGradeId, SeatGradeDto commandDto);
-}
