+++ showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/venue/seatgrade/DeleteSeatGradeUseCase.java
@@ -1,4 +1,4 @@
-package com.showhive.admin.application.command.usecase.seatgrade;
+package com.showhive.admin.application.command.usecase.venue.seatgrade;
 
 public interface DeleteSeatGradeUseCase {
     void deleteSeatGrade(long seatGradeId);
