+++ showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/venue/seatgrade/UpdateSeatGradeUseCase.java
@@ -1,0 +1,7 @@
+package com.showhive.admin.application.command.usecase.venue.seatgrade;
+
+import com.showhive.admin.application.command.dto.venue.SeatGradeDto;
+
+public interface UpdateSeatGradeUseCase {
+    void update(long seatGradeId, SeatGradeDto commandDto);
+}
