+++ showhive-boot/admin/src/main/java/com/showhive/admin/application/command/usecase/venue/seatgrade/impl/CreateSeatGradeUseCaseImpl.java
@@ -1,7 +1,7 @@
-package com.showhive.admin.application.command.usecase.seatgrade.impl;
+package com.showhive.admin.application.command.usecase.venue.seatgrade.impl;
 
-import com.showhive.admin.application.command.dto.SeatGradeDto;
-import com.showhive.admin.application.command.usecase.seatgrade.CreateSeatGradeUseCase;
+import com.showhive.admin.application.command.dto.venue.SeatGradeDto;
+import com.showhive.admin.application.command.usecase.venue.seatgrade.CreateSeatGradeUseCase;
 import com.showhive.venue.domain.SeatGrade;
 import com.showhive.venue.repository.command.SeatGradeCommandRepository;
 import lombok.RequiredArgsConstructor;
