+++ showhive-boot/admin/src/main/java/com/showhive/admin/interfaces/venue/dto/SeatGradeRequest.java
@@ -1,4 +1,4 @@
-package com.showhive.admin.interfaces.performance.dto;
+package com.showhive.admin.interfaces.venue.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import org.hibernate.validator.constraints.Length;
